pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brokerledger"
version = "0.1.0"
description = "Fully-local mortgage affordability analyser for UK brokers"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "BrokerLedger" }]
dependencies = [
    "PySide6~=6.7",
    "SQLAlchemy~=2.0",
    "pydantic~=2.7",
    "pydantic-settings~=2.3",
    "argon2-cffi~=23.1",
    "pdfplumber~=0.11",
    "pandas~=2.2",
    "openpyxl~=3.1",
    "rapidfuzz~=3.9",
    "httpx~=0.27",
    "python-dateutil~=2.9",
    "loguru~=0.7",
]

[project.optional-dependencies]
ocr = ["pytesseract~=0.3.10", "pdf2image~=1.17"]
encryption = ["pysqlcipher3~=1.2"]
dev = [
    "pytest~=8.2",
    "pytest-mock",
    "hypothesis~=6.100",
    "ruff~=0.5",
    "freezegun~=1.5",
]

[project.scripts]
brokerledger = "brokerledger.__main__:main"

[tool.setuptools.packages.find]
include = ["brokerledger*"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_ollama: requires a running Ollama instance (opt-in)",
]
